% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{bch_response_json}
\alias{bch_response_json}
\title{Parse a successful BCH response as JSON}
\usage{
bch_response_json(response)
}
\arguments{
\item{response}{A response returned by \code{\link[=bch_perform]{bch_perform()}}.}
}
\value{
Parsed JSON as nested R objects.
}
\description{
Parse a successful BCH response as JSON
}
\keyword{internal}
