% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{bch_result_accessors}
\alias{bch_result_accessors}
\alias{bch_data}
\alias{bch_metadata}
\alias{bch_provenance}
\alias{bch_problems}
\alias{bch_column_map}
\title{Extract a component from a BCH result}
\usage{
bch_data(x)

bch_metadata(x)

bch_provenance(x)

bch_problems(x)

bch_column_map(x)
}
\arguments{
\item{x}{A \code{bch_result} object.}
}
\value{
The requested tibble component.
}
\description{
Extract a component from a BCH result
}
