\name{bchR-package}
\alias{bchR}
\alias{bchR-package}
\title{bchR: Access to the Banco Central de Honduras Web API}
\description{
Provides a reproducible interface to the Banco Central de Honduras (BCH) Web API.
The initial package version focuses on secure API-key management, indicator-catalogue
retrieval, interactive catalogue browsing, and retrieval of individual indicator series.
}
\keyword{internal}
