% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{is_bch_result}
\alias{is_bch_result}
\title{Test whether an object is a BCH result}
\usage{
is_bch_result(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A single logical value.
}
\description{
Test whether an object is a BCH result
}
