% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{validate_bch_result}
\alias{validate_bch_result}
\title{Validate a BCH result}
\usage{
validate_bch_result(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
\code{x}, invisibly, or an error of class
\code{bch_invalid_result}.
}
\description{
Validate a BCH result
}
