% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bchR-package.R
\docType{package}
\name{bchR-package}
\alias{bchR}
\alias{bchR-package}
\title{bchR: Access to the Banco Central de Honduras Web API}
\description{
\code{bchR} provides a reproducible interface to the Banco Central de Honduras
(BCH) Web API. The initial package version focuses on secure API-key
management, indicator-catalogue retrieval, interactive catalogue browsing,
and retrieval of individual indicator series.
}
\author{
\strong{Maintainer}: Henry Osorto \email{henry.osorto@unah.edu.hn} (\href{https://orcid.org/0000-0002-4334-9179}{ORCID})

Authors:
\itemize{
  \item Henry Osorto \email{henry.osorto@unah.edu.hn} (\href{https://orcid.org/0000-0002-4334-9179}{ORCID})
}

}
\keyword{internal}
